requirements.txt
Appium-Python-Client==5.1.1
selenium>=4.26.0
pytest>=7.0.0
pytest-xdist>=3.0.0